pyproject.toml
[project]
name = "pytest-fly"
description = "pytest observer"
version = "0.0.22"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "James Abel", email = "j@abel.co" },
]
maintainers = [
    { name = "James Abel", email = "j@abel.co" },
]
license = { file = "LICENSE" }
classifiers = [
    "Framework :: Pytest",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Testing",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3 :: Only",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "pytest",
    "msqlite",
    "appdirs",
    "ismain",
    "numpy",
    "watchdog",
    "PySide6",
    "pref",
    "attrs",
    "typeguard",
    "balsa"
]
keywords = ["pytest", "observer", "testing"]

[project.urls]
Home = "https://github.com/jamesabel/pytest-fly"
Repository = "https://github.com/jamesabel/pytest-fly"
Documentation = "https://github.com/jamesabel/pytest-fly#readme"

[project.entry-points.pytest11]
fly = "pytest_fly.plugin"

[tool.black]
line-length = 192

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"